XSM: add Kconfig option XSM_ALLOW_BOOTLOADER_POLICY
Currently the built-in XSM policy only gets used if there is no other policy
provide by the bootloader. In this patch we add a Kconfig option that make this
optional to allow exlcusively using built-in policy during boot. This is
particularly important when booting Xen through the shim to ensure the XSM
policy gets measured and that it can't be replaced by another unmeasured policy
by the bootloader. Note that the XSM policy can still be updated after boot
(from dom0 for example) if the built-in policy allows it.

Signed-off-by: Tamas K Lengyel <lengyelt@ainfosec.com>
Acked-by: Daniel De Graaf <dgdegra@tycho.nsa.gov>
---
v2: Change Kconfig option name to XSM_ALLOW_BOOTLOADER_POLICY
    Make Kconfig option only depend on XSM_POLICY

Cc: Andrew Cooper <andrew.cooper3@citrix.com>
Cc: George Dunlap <George.Dunlap@eu.citrix.com>
Cc: Ian Jackson <ian.jackson@eu.citrix.com>
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Cc: Stefano Stabellini <sstabellini@kernel.org>
Cc: Tim Deegan <tim@xen.org>
Cc: Wei Liu <wei.liu2@citrix.com>
Cc: openxt@googlegroups.com