x86: Configure VMCS for Intel Processor Trace virtualization
This patch configure VMCS to make Intel Processor Trace
output address can be treat as guest physical address and
translated by EPT when ipt option is in guest mode.

Intel Processor Trace will be disabled in guest and the VMCS
configuration will be clear when part of the required
features is available in hardware or this feature is
disabled by user.

Signed-off-by: Luwei Kang <luwei.kang@intel.com>