x86/hvm: Pass IDT vectoring info through npfec
When a nested page-fault happens while delivering an event through the IDT due
to stack underflow, the IDT-vectoring information is not taken into account by
Xen. If the fault can't be resolved by lifting mem_access permissions or by
populting the entry from a parent VM, the MMIO emulation route is activated in
Xen. Trying to emulate the instruction that has such memory restrictions is
not something Xen can handle, throws X86EMUL_EXCEPTION, which leads to the
guest live-locking.

Fix it by crashing the guest to make it explicit that Xen can't handle such a
condition.

Signed-off-by: Tamas K Lengyel <tamas.lengyel@intel.com>